% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euc_sample.R
\docType{data}
\name{euc_sample}
\alias{euc_sample}
\title{Random 20 Numerical Values.}
\format{A data frame with 20 random variables:
\describe{
  \item{Value_One}{First Numerical Value}
  \item{Value_Two}{Second Numerical Value}

}}
\source{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\usage{
euc_sample
}
\description{
A dataset containing 20 numberical values for test Euclidean Algorithm.
}
\keyword{datasets}
