% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_func.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
Returns the greatest common divisor of \code{a} and \code{b}.
}
\description{
Euclid's algorithm, is an efficient method for computing the greatest common divisor (GCD) of two numbers, the largest number that divides both of them without leaving a remainder.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
\seealso{
https://en.wikipedia.org/wiki/Euclidean_algorithm for Euclidean algorithm
}
