% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Find the greatest common divisor of two numbers.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A Numerical value}

\item{b}{A Numerical Value}
}
\value{
Gratest common divior of given values
}
\description{
Find the greatest common divisor of two numbers.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
